requirements.txt
diffusers==0.24.0
gradio==4.5.0
opencv-python
opencv-python-headless
scikit-image
torch
torchvision
einops
accelerate==0.30.1
transformers==4.41.1
colorlog
cupy-cuda12x==12.3.0
av==13.1.0
gpustat
decord
bitsandbytes
xformers==0.0.28
decord
bitsandbytes
huggingface_hub<=0.23.2
colorama
plotly
Pillow
DateTime
matplotlib
plyfile
HTML4Vision
timm
tensorboardX
imgaug
iopath
imagecorruptions
mmcv